* Refactoring, introduced ReasonController as child class of MidiController to   prepare to eventually have a AbletonLiveController, as well
