+++ packages/midi/src/index.js
@@ -1,5 +1,6 @@
 module.exports = {
   MidiClock: require("./MidiClock"),
   MidiController: require("./MidiController"),
+  ReasonController: require("./ReasonController"),
   MidiInterface: require("./MidiInterface")
 };
